working (basic) implementation of ECS, entity has map of Components which can be accessed
