+++ bobo_engine/src/Bobo/ECS/Component.h
@@ -1,0 +1,14 @@
+#pragma once
+
+class Entity;
+	class Component
+	{
+	public:
+		virtual ~Component() {}
+		virtual void Awake() {}
+		virtual void Update() {}
+		virtual void FixedUpdate() {}
+
+	protected:
+		Entity* p_Parent;
+	};
